move quizlet helper to its own file
