+++ api/search.js
@@ -4,40 +4,7 @@ const request = require('request');
 const co = require('co');
 const config = require('../config');
 const logger = require('../logger');
-
-function parseQuizletSet(quizletID) {
-	const url = `https://api.quizlet.com/2.0/sets/${quizletID}?client_id=${config.quizlet.clientID}`;
-	return new Promise((resolve) => {
-		request({ url }, (error, res) => {
-			if (error) {
-				logger.error('Error when requesting set from Quizlet', { error, quizletID });
-				return resolve(error);
-			}
-			const quizletSet = JSON.parse(res.body);
-			if (!quizletSet.title) return resolve('');
-			const qset = { title: quizletSet.title, questions: [], privacy: false };
-			const answerPool = [];
-			quizletSet.terms.forEach((term) => {
-				if (term.definition.length > 0 && answerPool.indexOf(term.definition) < 0) {
-					answerPool.push(term.definition);
-				}
-			});
-			if (answerPool.length < 10) return resolve('');
-			quizletSet.terms.forEach((term, index) => {
-				const correctAnswer = term.definition;
-				const incorrectAnswers = [];
-				while (incorrectAnswers.length < 3) {
-					const randomAnswer = answerPool[Math.floor(Math.random() * answerPool.length)];
-					if (incorrectAnswers.indexOf(randomAnswer) < 0 && randomAnswer !== correctAnswer) {
-						incorrectAnswers.push({ text: randomAnswer, id: incorrectAnswers.length });
-					}
-				}
-				qset.questions.push({ text: term.term, correctAnswer, incorrectAnswers, id: index });
-			});
-			resolve(qset);
-		});
-	});
-}
+const quizlet = require('./helpers/quizlet');
 
 module.exports = function (req, res) {
 	if (!req.body.query) {
@@ -64,7 +31,7 @@ module.exports = function (req, res) {
 			return res.end(JSON.stringify(qsets));
 		}
 		const quizletSearchResults = JSON.parse(body);
-		const parsedSets = quizletSearchResults.sets.map((set) => parseQuizletSet(set.id));
+		const parsedSets = quizletSearchResults.sets.map((set) => quizlet.parseSet(set.id));
 		co(function* () {
 			return yield parsedSets;
 		}).then((sets) => {
